+++ src/bit_utils.rs
@@ -1,6 +1,5 @@
 use std::mem::size_of;
 use std::ops::ControlFlow;
-use std::any::TypeId;
 use crate::Primitive;
 
 /// Block ordering undefined. But same as [get_array_bit].
@@ -87,6 +86,7 @@ pub unsafe fn get_bit_unchecked<T: Primitive>(block: T, bit_index: usize) -> boo
 pub unsafe fn zero_high_bits_unchecked<T: Primitive>(block: T, bit_index: usize) -> T {
     #[cfg(target_feature = "bmi2")]
     {
+        use std::any::TypeId;
         if TypeId::of::<T>() == TypeId::of::<u64>(){
             return Primitive::from_u64(
                 std::arch::x86_64::_bzhi_u64(block.as_u64(), bit_index as u32)